start to move code to common
