+++ failgood/jvm/src/failgood/Failgood.jvm.kt
@@ -1,0 +1,5 @@
+package failgood
+
+import failgood.internal.execution.context.sourceInfo
+
+actual fun getSourceInfo(): failgood.SourceInfo = sourceInfo()
